mcp-server/requirements.txt
# Python MCP Server Dependencies
mcp>=1.1.0
supabase>=2.9.0

# AI Providers (Multi-provider support)
anthropic>=0.39.0
openai>=1.12.0
google-generativeai>=0.3.2
azure-ai-openai>=1.0.0

# OCR & Image Processing
pytesseract>=0.3.13
Pillow>=11.0.0

# Web Scraping (Async)
aiohttp>=3.9.1
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Utilities
python-dotenv>=1.0.0
